+++ debug_harness/Parser/ParseTree/tree_types.h
@@ -164,7 +164,7 @@ struct Expression
         } function_call_data;
 
         /* Integer number */
-        int i_num;
+        long long i_num;
 
         /* Real number */
         float r_num;
